refactor(web): extract MCP response header decoration into a lib
The MCP streamable route defined applyMcpHeaders() (and its CORS header map)
inline, so the CORS + security + no-store header application could not be
unit-tested without the handler. Move them into mcp-headers-lib.ts and import
the helper back.

- mcp-headers-lib.ts: add applyMcpHeaders(response) with the private
  mcpCorsHeaders map.
- routes/api/mcp.ts: import applyMcpHeaders; drop the local copy, the CORS map,
  and the now-unused applySecurityHeaders import.
- tests: cover CORS/no-store application, status/statusText/body preservation,
  and incoming header carry-through.

No user-facing or behavior change; MCP responses carry identical headers.
